--- nebula_ros/include/nebula_ros/hesai/hesai_hw_interface_ros_wrapper.hpp
@@ -1,100 +1,0 @@
-#ifndef NEBULA_HesaiHwInterfaceRosWrapper_H
-#define NEBULA_HesaiHwInterfaceRosWrapper_H
-
-#include "nebula_common/hesai/hesai_common.hpp"
-#include "nebula_common/nebula_common.hpp"
-#include "nebula_hw_interfaces/nebula_hw_interfaces_hesai/hesai_hw_interface.hpp"
-#include "nebula_ros/common/nebula_hw_interface_ros_wrapper_base.hpp"
-#include "boost_tcp_driver/tcp_driver.hpp"
-
-#include <ament_index_cpp/get_package_prefix.hpp>
-#include <rclcpp/rclcpp.hpp>
-#include <rclcpp_components/register_node_macro.hpp>
-
-#include "pandar_msgs/msg/pandar_packet.hpp"
-#include "pandar_msgs/msg/pandar_scan.hpp"
-
-#include <boost/asio.hpp>
-
-#include <mutex>
-#include <thread>
-
-namespace nebula
-{
-namespace ros
-{
-/// @brief Get parameter from rclcpp::Parameter
-/// @tparam T
-/// @param p Parameter from rclcpp parameter callback
-/// @param name Target parameter name
-/// @param value Corresponding value
-/// @return Whether the target name existed
-template <typename T>
-bool get_param(const std::vector<rclcpp::Parameter> & p, const std::string & name, T & value)
-{
-  auto it = std::find_if(p.cbegin(), p.cend(), [&name](const rclcpp::Parameter & parameter) {
-    return parameter.get_name() == name;
-  });
-  if (it != p.cend()) {
-    value = it->template get_value<T>();
-    return true;
-  }
-  return false;
-}
-
-/// @brief Hardware interface ros wrapper of hesai driver
-class HesaiHwInterfaceRosWrapper final : public rclcpp::Node, NebulaHwInterfaceWrapperBase
-{
-  drivers::HesaiHwInterface hw_interface_;
-  Status interface_status_;
-
-  drivers::HesaiSensorConfiguration sensor_configuration_;
-
-  /// @brief Received Hesai message publisher
-  rclcpp::Publisher<pandar_msgs::msg::PandarScan>::SharedPtr pandar_scan_pub_;
-
-  /// @brief Initializing hardware interface ros wrapper
-  /// @param sensor_configuration SensorConfiguration for this driver
-  /// @return Resulting status
-  Status InitializeHwInterface(
-    const drivers::SensorConfigurationBase & sensor_configuration) override;
-  /// @brief Callback for receiving PandarScan
-  /// @param scan_buffer Received PandarScan
-  void ReceiveScanDataCallback(std::unique_ptr<pandar_msgs::msg::PandarScan> scan_buffer);
-
-public:
-  explicit HesaiHwInterfaceRosWrapper(const rclcpp::NodeOptions & options);
-  ~HesaiHwInterfaceRosWrapper() noexcept override;
-  /// @brief Start point cloud streaming (Call SensorInterfaceStart of HwInterface)
-  /// @return Resulting status
-  Status StreamStart() override;
-  /// @brief Stop point cloud streaming (not used)
-  /// @return Resulting status
-  Status StreamStop() override;
-  /// @brief Shutdown (not used)
-  /// @return Resulting status
-  Status Shutdown() override;
-  /// @brief Get configurations from ros parameters
-  /// @param sensor_configuration Output of SensorConfiguration
-  /// @return Resulting status
-  Status GetParameters(drivers::HesaiSensorConfiguration & sensor_configuration);
-
-private:
-  uint16_t delay_hw_ms_;
-  bool retry_hw_;
-  std::mutex mtx_config_;
-  OnSetParametersCallbackHandle::SharedPtr set_param_res_;
-  /// @brief rclcpp parameter callback
-  /// @param parameters Received parameters
-  /// @return SetParametersResult
-  rcl_interfaces::msg::SetParametersResult paramCallback(
-    const std::vector<rclcpp::Parameter> & parameters);
-  /// @brief Updating rclcpp parameter
-  /// @return SetParametersResult
-  std::vector<rcl_interfaces::msg::SetParametersResult> updateParameters();
-};
-
-}  // namespace ros
-}  // namespace nebula
-
-#endif  // NEBULA_HesaiHwInterfaceRosWrapper_H
